more implementations, solution_v2 was accepted

+++ leetcode/leetcode_number-of-1-bits/Solution_v1.java
@@ -2,7 +2,7 @@ public class Solution_v1{
 
 	public int hammingWeight(int n) {
 		int count = 0;
-	 	for(int i = 1; i<=n; i*=2){
+	 	for(long i = 1; i<=n; i*=2){
 	 		System.out.println("i is now " + i);
 	 		if((i & n) != 0){
 	 			count++;
